Implement redimap script, with crude polling.
Execute `redimap` to read mailboxes and do magic. This can be executed
from a cron. Alternatively, for crude polling, call with `1` as the
first argument, such as `redimap 1`. This will sleep for
polling_interval seconds after one pass through the loop.

Note that polling_interval=60 is not the same as it being run every
minute. Indeedy.