Restore ability to set config from ENV vars.
On reflection, I have repented. However, the change to how ENV vars
were read previously is that they are expected to be lowercase, in
contrast to popular implementation. But read

    http://pubs.opengroup.org/onlinepubs/007908799/xbd/envvar.html

So, for the executable, each config setting is now set in the following
order of priority, with a mixture of methods possible:

- command-line parameters (top priority; overrides others)
- ENV vars
- defaults

Note that when using Redimap as a library, only the configure block is
used; there is no reading of ENV vars, and of course no reading of
command-line parameters.