Revert "Restore ability to set config from ENV vars."
My apologies, world; all this time this song is best. The aim of
allowing Redimap config to be set directly via ENV vars was to make it
easier to work as a service as part of a larger application (such as in
a Procfile). However, usage indicated that many things expect ENV vars
to be uppercased, so the ENV vars didn't turn out to be as useful as
hoped. In addition, it really is rather easy to set ENV vars, if you
so desire, using the command-line parameters, such as `-h $IMAP_HOST`.
Supporting ENV vars directly adds bloat to the code.

So, no direct ENV var support for Redimap. If including as a library,
use a configure block. If running as a service, specify command-line
parameters. If you'd like to use ENV vars, simply pass those as
command-line parameters.

This reverts commit 16daf880c83dcb63d42b395a4e7513f3d159cdbc.